Separate wasm sources for vanilla JS and Angular apps
